+++ src/collective/dms/mailcontent/browser/utils.py
@@ -1,5 +1,6 @@
 # encoding: utf-8
 
+from collective.dms.mailcontent.dmsmail import IDmsIncomingMail
 from Products.Five import BrowserView
 
 
@@ -8,6 +9,6 @@ class UtilsMethods(BrowserView):
 
     def outgoingmail_folder(self):
         """ Get a folder for outgoing mail """
-        if self.context.portal_type == 'dmsincomingmail':
+        if IDmsIncomingMail.providedBy(self.context):
             return self.context.__parent__
         return self.context