+++ CHANGES.rst
@@ -9,7 +9,7 @@ Changelog
   [sgeulette]
 - RelatedDocs field (reply_to) use now object_provides criteria
   [sgeulette]
-- Replaced dmsincomingmail portal_type test by IDmsIncomingMail.providedBy
+- Replaced portal_type test by schema interface providedBy test
   [sgeulette]
 
 1.4.3 (2020-10-07)